made new test for the files
