+++ simple_package/statistics.py
@@ -22,7 +22,6 @@
 ##    numpy and matplotlib packages are not installed.
 
 import numpy as np
-import matplotlib.pyplot as plt
 
 def calculate_statistics(data):
     if not isinstance(data, (np.ndarray, list)):
@@ -40,26 +39,3 @@ def display_statistics(mean, median, std_dev):
     print(f"Mean: {mean}")
     print(f"Median: {median}")
     print(f"Standard Deviation: {std_dev}")
-
-def plot_histogram(data, mean, median):
-    plt.hist(data, bins=50, alpha=1, label='Data')
-    plt.axvline(mean, color='r', label='Mean')
-    plt.axvline(median, color='g', label='Median')
-    plt.legend()
-    plt.show()
-
-''' added a test to make sure code works as desired '''
-def main():
-    try:
-        import numpy as np
-        import matplotlib.pyplot as plt
-    except ImportError:
-        raise ImportError("numpy and/or matplotlib packages are not installed")
-    
-    data = np.array([1, 2, 3, 4, 5, 5, 3, 1, 4, 6, 8, 3, 10])  # example data
-    mean, median, std_dev = calculate_statistics(data)
-    display_statistics(mean, median, std_dev)
-    plot_histogram(data, mean, median)
-
-if __name__ == "__main__":
-    main()
